Dostosowanie API do filtrów i refactor autoryzacji (#11)
* Dostosowanie api pod filtry

* Dostosowanie endpointa z kursami personalnymi do filtrów

* Dostosowanie middleware multera

* Refactor systemu autoryzacji

+++ src/controllers/course.ts
@@ -82,7 +82,7 @@ export const updateCourse = (req: Request, res: Response) => {
 
             coursesDatabase.query(
                 'UPDATE Courses SET trainer_id = ?, title = ?, description = ?, language = ?, level = ?, location = ?, image_path = ? WHERE id = ?',
-                [data.trainer_id, data.title, data.description, data.language, data.level, data.location, req.file?.path ? req.file.path : originalImagePath, courseId],
+                [data.trainer_id, data.title, data.description, data.language, data.level, data.location, req.file?.path ? req.file.path : (originalImagePath ?? null), courseId],
                 (error, result) => {
                     if (error) {
                         if (req.file?.path)
@@ -110,7 +110,7 @@ export const updateCourse = (req: Request, res: Response) => {
                                 }
                             )
                         })
-                        if (req.file?.path)
+                        if (req.file?.path && originalImagePath)
                             fs.unlink(originalImagePath, error => console.log(error));
                         return res.json({
                             message: 'Udało się zaktualizować kurs!'
@@ -165,8 +165,6 @@ export const getAllCourses = (req: Request, res: Response) => {
                         }]
                     })
                     return res.json(data);
-
-                    // console.log(sortLessons(result));
                 })
             })
         })
@@ -181,18 +179,27 @@ export const getPersonalCourses = (req: Request, res: Response) => {
                     if (error)
                         return res.sendStatus(500);
 
-                    result.forEach((course: any) => {
-                        const isEnrolled = !!bindedCoursesData.find((elem: any) => elem.course_id === course.id && elem.enroll_id === userId);
-                        const isOwner = !!bindedCoursesData.find((elem: any) => elem.course_id === course.id && elem.owner_id === userId);
-                        if (isEnrolled || isOwner) {
-                            data = [...data, {
-                                ...course,
-                                isEnrolled: isEnrolled,
-                                isOwner: isOwner
-                            }]
-                        }
+                    coursesDatabase.query('SELECT Lessons.course_id, Lessons.date, Lessons.timeStart, Lessons.timeFinish FROM Lessons', [], (error, lessons) => {
+                        if (error)
+                            return res.sendStatus(500);
+
+                        const sortedLessons = sortDatabaseLessons(lessons);
+                        result.forEach((course: any) => {
+                            const isEnrolled = !!bindedCoursesData.find((elem: any) => elem.course_id === course.id && elem.enroll_id === userId);
+                            const isOwner = !!bindedCoursesData.find((elem: any) => elem.course_id === course.id && elem.owner_id === userId);
+                            if (isEnrolled || isOwner) {
+                                const filteredLessons = sortedLessons.filter(lesson => lesson.course_id === course.id);
+                                data = [...data, {
+                                    ...course,
+                                    isEnrolled: isEnrolled,
+                                    isOwner: isOwner,
+                                    firstLesson: filteredLessons?.[0],
+                                    lastLesson: filteredLessons?.[filteredLessons.length - 1]
+                                }]
+                            }
+                        })
+                        return res.json(data);
                     })
-                    return res.json(data);
                 })
         })
 }
